backend/pyproject.toml
[project]
name = "wayfr-backend"
version = "0.1.0"
description = "wayfr real-time AI navigation backend"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "websockets>=13.1",

    # AI / ML
    "google-generativeai>=0.8.0",
    "replicate>=0.34.0",
    "google-cloud-vision>=3.8.0",
    "gtts>=2.5.0",

    # Database
    "supabase>=2.10.0",
    "upstash-redis>=1.3.0",

    # 3D / numeric / video
    "numpy>=1.26.0",
    "pillow>=10.4.0",
    "opencv-python-headless>=4.8.0",
    "python-geohash>=0.8.5",
    "trimesh>=4.0.0",
    "scipy>=1.13.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",

    # Utils
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "structlog>=24.4.0",

    # Modal (serverless GPU)
    "modal>=0.68.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.32.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "db", "ml", "models", "services"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
